Tudo pronto

+++ Desafio/API_movieRent/src/modules/movies/services/DeleteMovieService.ts
@@ -1,3 +1,4 @@
+import { AppError } from "../../../shared/errors/AppError";
 import { IMovieRepository } from "../domain/repositories-models/IMovieRepository";
 
 
@@ -6,6 +7,9 @@ export default class DeleteMovieService{
 
   public async execute(id:string){
     const movie = await this.repository.findById(id);
+    if(!movie){
+      throw new AppError('filme não cadastrado',404);
+    }
     if (movie) await this.repository.remove(movie);
